requirements.txt
Django==4.2.7
djangorestframework==3.14.0
jwt==1.3.1
PyJWT==2.8.0
